requirements.txt
streamlit==1.18.1
plotly
altair==4.0
python-dotenv